src/Text2SQL/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
mariadb>=1.1.0
anthropic>=0.7.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.0.0 